fixes with table
(cherry picked from commit 6d918aa4627cdd9f836a9ab25c5f04445f7edb73)
