+++ src/astro/dataframe/__init__.py
@@ -22,18 +22,10 @@ def dataframe(
     }
     if task_id:
         param_map["task_id"] = task_id
-    df_class = SqlDataframeOperator
-    if kwargs.get("_experimental", False):
-        from astro.sql.operators.sql_dataframe_refactor import (
-            SqlDataframeOperator as DFNew,
-        )
-
-        df_class = DFNew
-        kwargs.pop("_experimental")
     decorated_function: Callable[..., pd.DataFrame] = task_decorator_factory(
         python_callable=python_callable,
         multiple_outputs=multiple_outputs,
-        decorated_operator_class=df_class,  # type: ignore
+        decorated_operator_class=SqlDataframeOperator,  # type: ignore
         **param_map,
     )
     return decorated_function
