+++ src/astro/sql/operators/agnostic_sql_merge.py
@@ -3,7 +3,7 @@ from typing import Dict, List, Union
 from airflow.exceptions import AirflowException
 
 from astro.constants import Database
-from astro.sql.operators.sql_decorator import SqlDecoratedOperator
+from astro.sql.operators.transform import TransformOperator
 from astro.sql.table import Table
 from astro.utils.bigquery_merge_func import bigquery_merge_func
 from astro.utils.database import create_database_from_conn_id
@@ -17,7 +17,7 @@ from astro.utils.sqlite_merge_func import sqlite_merge_func
 from astro.utils.task_id_helper import get_unique_task_id
 
 
-class SqlMergeOperator(SqlDecoratedOperator):
+class SqlMergeOperator(TransformOperator):
     template_fields = ("target_table", "merge_table")
 
     def __init__(
